@@ -44,52 +44,55 @@ ActiveAdmin.register Equipment::Radius::AccountingProfile do
   form do |f|
     f.semantic_errors *f.object.errors.keys.uniq
     f.inputs form_title do
-      f.input :name
-      f.input :server
-      f.input :port
-      f.input :secret
+      f.input :name, hint: I18n.t('hints.equipment.radius_accounting_profile.name')
+      f.input :server, hint: I18n.t('hints.equipment.radius_accounting_profile.server')
+      f.input :port, hint: I18n.t('hints.equipment.radius_accounting_profile.port')
+      f.input :secret, hint: I18n.t('hints.equipment.radius_accounting_profile.secret')
       f.input :timeout, hint: "Request timeout in ms. Allowed values #{Equipment::Radius::AccountingProfile::TIMEOUT_MIN}..#{Equipment::Radius::AccountingProfile::TIMEOUT_MAX}"
       f.input :attempts, hint: "Max request attempts. Allowed values #{Equipment::Radius::AccountingProfile::ATTEMPTS_MIN}..#{Equipment::Radius::AccountingProfile::ATTEMPTS_MAX}"
       f.input :enable_start_accounting
       f.input :enable_interim_accounting
-      f.input :interim_accounting_interval
+      f.input :interim_accounting_interval, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_accounting_interval')
       f.input :enable_stop_accounting
     end
     f.inputs "Start packet attributes" do
       f.has_many :start_avps do |t|
-        t.input :type_id, hint: "Attribute type, see rfc2865"
-        t.input :name, hint: "Informational only"
+        t.input :type_id, hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.type_id')
+        t.input :name, hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.name')
         t.input :is_vsa
-        t.input :vsa_vendor_id
-        t.input :vsa_vendor_type
-        t.input :value
-        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS
+        t.input :vsa_vendor_id, hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.vsa_vendor_id')
+        t.input :vsa_vendor_type, hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.vsa_vendor_type')
+        t.input :value, hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.value')
+        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS,
+                hint: I18n.t('hints.equipment.radius_accounting_profile.start_avps.format')
         t.input :_destroy, as: :boolean, required: false, label: 'Remove' unless  t.object.new_record?
       end
     end
 
     f.inputs "Interim packet attributes" do
       f.has_many :interim_avps do |t|
-        t.input :type_id, hint: "Attribute type, see rfc2865"
-        t.input :name, hint: "Informational only"
+        t.input :type_id, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.type_id')
+        t.input :name, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.name')
         t.input :is_vsa
-        t.input :vsa_vendor_id
-        t.input :vsa_vendor_type
-        t.input :value
-        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS
+        t.input :vsa_vendor_id, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.vsa_vendor_id')
+        t.input :vsa_vendor_type, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.vsa_vendor_type')
+        t.input :value, hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.value')
+        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS,
+                hint: I18n.t('hints.equipment.radius_accounting_profile.interim_avps.format')
         t.input :_destroy, as: :boolean, required: false, label: 'Remove' unless  t.object.new_record?
       end
     end
 
     f.inputs "Stop packet attributes" do
       f.has_many :stop_avps do |t|
-        t.input :type_id, hint: "Attribute type, see rfc2865"
-        t.input :name, hint: "Informational only"
+        t.input :type_id, hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.type_id')
+        t.input :name, hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.name')
         t.input :is_vsa
-        t.input :vsa_vendor_id
-        t.input :vsa_vendor_type
-        t.input :value
-        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS
+        t.input :vsa_vendor_id, hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.vsa_vendor_id')
+        t.input :vsa_vendor_type, hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.vsa_vendor_type')
+        t.input :value, hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.value')
+        t.input :format, as: :select, collection: Equipment::Radius::Attribute::FORMATS,
+                hint: I18n.t('hints.equipment.radius_accounting_profile.stop_avps.format')
         t.input :_destroy, as: :boolean, required: false, label: 'Remove' unless  t.object.new_record?
       end
     end
