@@ -31,9 +31,11 @@ ActiveAdmin.register Report::CustomerTraffic, as: 'CustomerTraffic' do
 
   form do |f|
     f.inputs do
-      f.input :date_start, as: :date_time_picker, wrapper_html: { class: 'datetime_preset_pair', data: { show_time: 'true' } }
-      f.input :date_end, as: :date_time_picker
-      f.input :customer, as: :select, input_html: {class: 'chosen'}, collection: Contractor.where(customer: true)
+      f.input :date_start, as: :date_time_picker, wrapper_html: { class: 'datetime_preset_pair', data: { show_time: 'true' } },
+              hint: I18n.t('hints.reports.customer_traffic.date_start')
+      f.input :date_end, as: :date_time_picker, hint: I18n.t('hints.reports.customer_traffic.date_end')
+      f.input :customer, as: :select, input_html: {class: 'chosen'}, collection: Contractor.where(customer: true),
+              hint: I18n.t('hints.reports.customer_traffic.customer')
       f.input :send_to, as: :select, input_html: {class: 'chosen', multiple: true}, collection: Billing::Contact.collection, hint: f.object.send_to_hint
     end
     f.actions
