@@ -30,8 +30,9 @@ ActiveAdmin.register Report::CustomerTrafficScheduler, as: 'CustomerTrafficSched
 
   form do |f|
     f.inputs do
-      f.input :period
-      f.input :customer, as: :select, input_html: {class: 'chosen'}, collection: Contractor.where(customer: true)
+      f.input :period, hint: I18n.t('hints.reports.customer_traffic_scheduler.period')
+      f.input :customer, as: :select, input_html: {class: 'chosen'}, collection: Contractor.where(customer: true),
+              hint: I18n.t('hints.reports.custom_cdr.customer')
       f.input :send_to, as: :select, input_html: {class: 'chosen', multiple: true}, collection: Billing::Contact.collection, hint: f.object.send_to_hint
     end
     f.actions
