@@ -27,10 +27,11 @@ ActiveAdmin.register Routing::RateplanDuplicator do
   form do |f|
     f.semantic_errors *f.object.errors.keys
     f.inputs "Copy rateplan" do
-      f.input :id, as: :hidden
-      f.input :name
-      f.input :profit_control_mode_id, collection: Routing::RateProfitControlMode.all
-      f.input :send_quality_alarms_to, as: :select, input_html: {class: 'chosen-sortable', multiple: true}, collection: Billing::Contact.collection
+      f.input :id, as: :hidden, hint: I18n.t('hints.routing.rateplan_duplicator.id')
+      f.input :name, hint: I18n.t('hints.routing.rateplan_duplicator.name')
+      f.input :profit_control_mode_id, collection: Routing::RateProfitControlMode.all, hint: I18n.t('hints.routing.rateplan_duplicator.profit_control_mode_id')
+      f.input :send_quality_alarms_to, as: :select, input_html: {class: 'chosen-sortable', multiple: true}, collection: Billing::Contact.collection,
+              hint: I18n.t('hints.routing.rateplan_duplicator.send_quality_alarms_to')
     end
     f.actions do
       action(:submit)
