@@ -39,7 +39,7 @@ class ApplicationController < ActionController::Base
 
 
   def user_for_paper_trail
-    admin_user_signed_in? ? current_admin_user : 'Unknown user'
+    admin_user_signed_in? ? current_admin_user.try(:id) : 'Unknown user'
   end
 
   def current_ability
