@@ -1,5 +1,6 @@
 FactoryGirl.define do
   factory :rateplan, class: Rateplan do
-    name nil
+    sequence(:name) { |n| "rateplan#{n}" }
+    association :profit_control_mode, factory: :rate_profit_control_mode
   end
 end
