@@ -2,7 +2,7 @@
 ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
-require 'capybara/rspec'
+# require 'capybara/rspec'
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
@@ -41,7 +41,10 @@ RSpec.configure do |config|
       :codecs,
       :dump_level,
       'class4.dtmf_send_modes',
-      'class4.dtmf_receive_modes'
+      'class4.dtmf_receive_modes',
+      'class4.gateway_rel100_modes',
+      'class4.transport_protocols',
+      'sys.timezones'
   ]
 
   # RSpec Rails can automatically mix in different behaviours to your tests
@@ -80,4 +83,8 @@ RSpec.configure do |config|
   config.after(:each) do
     DatabaseCleaner.clean
   end
+end
+
+RspecApiDocumentation.configure do |config|
+  config.format = :json
 end
